// src/app/interfaces/planta-es.interface.ts
export interface PlantaEs {
  id: number;
  id_planta: number;
  nombre_comun: string;
  nombre_cientifico: string;
  familia: string;
  origen: string;
  tipo: string;
  ciclo: string;
  riego: string;
  luz_solar: string;
  nivel_cuidado: string;
  velocidad_crecimiento: string;
  nivel_mantenimiento: string;
  temporada_floracion: string;
  color_flor: string;
  color_hojas: string;
  altura: number;
  unidad_altura: string;
  extension: number;
  unidad_extension: string;
  descripcion: string;
  suelo: string;
  propagacion: string;
  resistencia_min: number;
  resistencia_max: number;
  interior: boolean;
  venenosa_humanos: boolean;
  venenosa_mascotas: boolean;
  fruto_comestible: boolean;
  temporada_cosecha: string;
  valor_nutricional_fruto: string;
  atrae: string;
  susceptibilidad_plagas: string;
  tolerante_sequia: boolean;
  flores: boolean;
  mapa_ubicacion_resistencia: string;
  fecha_creacion: string;
}